added routes and implemented book future line numbers

+++ ITD/src/main/java/com/example/demo/model/entities/WorkingHour.java
@@ -16,11 +16,11 @@ public class WorkingHour {
 
     @Column(name = "openingTime")
     @Expose
-    private long from;
+    private int from;
 
     @Column(name = "closingTime")
     @Expose
-    private long until;
+    private int until;
 
     @OneToOne(mappedBy = "workingHour")
     private Store store;
@@ -33,19 +33,19 @@ public class WorkingHour {
         this.id = id;
     }
 
-    public long getFrom() {
+    public int getFrom() {
         return from;
     }
 
-    public void setFrom(long from) {
+    public void setFrom(int from) {
         this.from = from;
     }
 
-    public long getUntil() {
+    public int getUntil() {
         return until;
     }
 
-    public void setUntil(long until) {
+    public void setUntil(int until) {
         this.until = until;
     }
